use types from sgb-v / sgb-xi
